# go-testing/go-web-testing/docs/swagger.yaml
definitions:
...
  web.response:
    properties:
      data: {}
    type: object
info:
  contact:
    name: API Support
    url: https://localhost:8080/
  description: This API Handle MELI Products.
  termsOfService: https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones
  title: MELI Bootcamp API FP
  version: "1.0"
paths:
...
  /api/v1/warehouses:
    delete:
      consumes:
      - application/json
      description: delete warehouses
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
      summary: Delete warehouse
      tags:
      - Warehouses
    get:
      consumes:
      - application/json
      description: get warehouses
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
      summary: Get warehouse by id
      tags:
      - Warehouses
    patch:
      consumes:
      - application/json
      description: update warehouses
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
      summary: Update warehouse
      tags:
      - Warehouses
    post:
      consumes:
      - application/json
      description: create warehouses
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
      summary: Create warehouse
      tags:
      - Warehouses
...
swagger: "2.0"